+++ src/components/FormatStatsRender.tsx
@@ -1,11 +1,11 @@
 import { FC } from "react";
 import { useMemo } from "react";
 import { RepositoryContributionsCard } from "./RepositoryContributionsCard";
+import { ExportDropdown } from "./ExportDropdown";
 import {
   PullRequestContributionsByRepository,
   RepositoryRenderFormat,
 } from "@/types/github";
-import ExportDropdown from "./ExportDropdown";
 import { exportAsJSON, exportAsText, generateText } from "@/utils";
 
 type NoContributionsProps = {
