+++ src/components/index.ts
@@ -1,4 +1,5 @@
 export * from "./ReposFilters";
+export * from "./ExportDropdown";
 export * from "./FormatStatsRender";
 export * from "./RepositoryContributionsCard";
 export * from "./CardSkeleton";
