+++ src/utils/exportAsText.ts
@@ -1,0 +1,5 @@
+import { downloadBlob } from ".";
+
+export const exportAsText = (text: string) => {
+  downloadBlob(new Blob([text], { type: "text/plain" }), "data.txt");
+};
