+++ src/utils/index.ts
@@ -1,3 +1,5 @@
+export * from "./exportAsImage";
+export * from "./exportAsJSON";
+export * from "./exportAsText";
 export * from "./downloadBlob";
 export * from "./generateText";
-export * from "./exportRepositories";
